requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pandas==2.1.3
openpyxl==3.1.2
sqlalchemy==2.0.23
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0